Moved functions in utils.js
